+++ spring-cloud-tencent-plugin-starters/spring-cloud-tencent-lossless-plugin/src/main/java/com/tencent/cloud/plugin/lossless/SpringCloudLosslessActionProvider.java
@@ -22,11 +22,10 @@ import java.util.Map;
 
 import com.tencent.cloud.common.util.OkHttpUtil;
 import com.tencent.cloud.plugin.lossless.config.LosslessProperties;
-import com.tencent.cloud.plugin.lossless.transfomer.DiscoveryNamespaceGetter;
+import com.tencent.cloud.rpc.enhancement.transformer.RegistrationTransformer;
 import com.tencent.polaris.api.plugin.lossless.InstanceProperties;
 import com.tencent.polaris.api.plugin.lossless.LosslessActionProvider;
 import com.tencent.polaris.api.pojo.BaseInstance;
-import com.tencent.polaris.api.pojo.DefaultBaseInstance;
 import com.tencent.polaris.api.utils.StringUtils;
 
 import org.springframework.cloud.client.serviceregistry.Registration;
@@ -89,20 +88,8 @@ public class SpringCloudLosslessActionProvider implements LosslessActionProvider
 				losslessProperties.getHealthCheckPath(), headers);
 	}
 
-	public static BaseInstance getBaseInstance(Registration registration, DiscoveryNamespaceGetter namespaceGetter) {
-		return getBaseInstance(registration, registration.getPort(), namespaceGetter);
+	public static BaseInstance getBaseInstance(Registration registration, RegistrationTransformer registrationTransformer) {
+		return registrationTransformer.transform(registration);
 	}
 
-	public static BaseInstance getBaseInstance(Registration registration, Integer port,
-												DiscoveryNamespaceGetter namespaceGetter) {
-		DefaultBaseInstance baseInstance = new DefaultBaseInstance();
-		if (namespaceGetter != null) {
-			baseInstance.setNamespace(namespaceGetter.getNamespace());
-		}
-		baseInstance.setService(registration.getServiceId());
-		// before web start, port in registration not init
-		baseInstance.setPort(port);
-		baseInstance.setHost(registration.getHost());
-		return baseInstance;
-	}
 }