+++ CHANGELOG.md
@@ -9,3 +9,4 @@
 - [feat:add polaris ThreadLocal plugin.](https://github.com/Tencent/spring-cloud-tencent/pull/1255)
 - [feat:upgrade jackson version.](https://github.com/Tencent/spring-cloud-tencent/pull/1257)
 - [fix:fix wrong report when using Zuul with instance not found exception.](https://github.com/Tencent/spring-cloud-tencent/pull/1283)
+- [feat: merge lane router and lossless features from 2023](https://github.com/Tencent/spring-cloud-tencent/pull/1288)